refactor(#792): split block-mod lib_tests into model + registry suites (sweep)
lib_tests.rs 694 -> 464, + lib_registry_tests.rs 240 (both under cap; off debt).
The native registry-level process tests + display-name/brand/type_label helpers move
to the sibling; default_params becomes pub(super), imported via super::tests. block-mod
tests pass, zero warnings.
